Minor text changes in elements.html
Just some text changes where the text was either broken or was a bit off.
i.e. "Use class x to <button>" isn't even valid english, the correct wording would either add x to or use x with. As most of the text already has the Add x to wording I opted for that.