Add dropdown functionality when hover or focus
I don't understand the functionality of the .on-focus class. While the docs say 'You may add the is-focused class to form-autocomplete-input to make it appear as focus state' it actually only manages shadow and border color.

I added an .autocomplete-dropdown class that hides the menu and shows it on hover or on focus. There is also a translate reset to prevent .menu hiding when mouse hovers the gap between .form-autocmplete-input and .menu