Add top border to tables that do not have a <thead> element
Consider a table without a `thead`, such as:

```html
<table class="table">
  <tbody>
    <tr>
      <td>The Shawshank Redemption</td>
      <td>Crime, Drama</td>
      <td>14 October 1994</td>
    </tr>
    <tr>
      <td>The Godfather</td>
      <td>Crime, Drama</td>
      <td>24 March 1972</td>
    </tr>
    <tr>
      <td>Schindler's List</td>
      <td>Biography, Drama, History</td>
      <td>4 February 1994</td>
    </tr>
    <tr>
      <td>Se7en</td>
      <td>Crime, Drama, Mystery</td>
      <td>22 September 1995</td>
    </tr>
  </tbody>
</table>
```

In this case, Spectre will not add a border to the top row, as the
corresponding CSS only defines bottom borders.

This fix adds the top border if there is no `thead` element (i.e.
`tbody` is the first child of the table).

Signed-off-by: Juan Cruz Viotti <jv@jviotti.com>