Fix form-select color
When using a dark theme with a light color, the select field will also have this color. Other input fields do not have this issue, as they use `color: $body-font-color;` instead. I've adapted that style.

This was initial set in the last version, so I am wondering, why this change was made:
https://github.com/picturepan2/spectre/commit/4b7f4493bb078f70236345ff045be9da34629ce9